+++ .gitignore
@@ -31,6 +31,6 @@ yarn-debug.log*
 .yarn-integrity
 .Ds_Store
 
-/app/controllers/client_secrets.json
+/client_secrets.json
 /token.yaml
 .credentials.json
