+++ app/assets/stylesheets/shared/comment_forms.css
@@ -63,7 +63,7 @@
   margin-bottom: 20px;
   padding-left: 30px;
   height: 7vh;
-  font-size: 40px;
+  font-size: 30px;
   max-width: 80vw;
   width: 50vw;
 }
