+++ app/assets/stylesheets/user/index.css
@@ -1,3 +1,7 @@
+.user-index-main {
+  background-color: aliceblue;
+}
+
 .index-user-dates-title {
   padding-top: 20px;
   text-align: center;
@@ -14,11 +18,13 @@
 .date-column-tables {
   display: flex;
   font-size: 20px;
+  background-color: whitesmoke;
 }
 
 .users-date-tables {
   display: flex;
   font-size: 20px;
+  background-color: whitesmoke;
 }
 
 .users-date-name-th {
