+++ app/assets/stylesheets/user/new.css
@@ -1,18 +1,20 @@
 .new-registration-board {
   width: 100vw;
   height: 80vh;
+  background-color: aliceblue;
 }
 
 .new-user-registration-title {
-  margin-top: 200px;
+  padding-top: 50px;
   text-align: center;
   font-size: 50px;
   font-weight: bolder;
+  
 }
 
 .mail-registration-link {
   margin-bottom: 50px;
-  margin-top: 200px;
+  margin-top: 100px;
   justify-content: center;
   text-align: center;
 }
