+++ app/assets/stylesheets/users.scss
@@ -56,3 +56,7 @@
   margin: 40px 0 20px;
   text-align: center;
 }
+
+#g-button {
+  margin-left: 30px;
+}
