--- app/helpers/gcalendars_helper.rb
@@ -1,2 +1,0 @@
-module GcalendarsHelper
-end
