+++ app/views/users/new.html.erb
@@ -15,4 +15,5 @@
       </div>
     <% end %>
   </div>
+</div>
 <%= render "shared/second-footer"%>
