--- spec/requests/gcalendars_request_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Gcalendars", type: :request do
-
-end
